Criação do Banco, Migrations, alteração nas PK e criação do script

+++ HBSIS_Padawan.Sistema.Boletim.Models/CursoMateria.cs
@@ -4,11 +4,11 @@ namespace HBSIS_Padawan.Sistema.Boletim.Models
 {
     public class CursoMateria
     {
-        public int Id { get; set; }
+        public long Id { get; set; }        
         public virtual Curso Curso { get; set; }
         public virtual Materia Materia { get; set; }
-        public int CursoId { get; set; }
-        public int MateriaId { get; set; }
+        public long CursoId { get; set; }
+        public long MateriaId { get; set; }
         public ICollection<CursoMateriaAluno> NotasAlunos { get; set; } = new List<CursoMateriaAluno>();
     }
 }
