+++ HBSIS_Padawan.Sistema.Boletim.Models/Materia.cs
@@ -4,10 +4,10 @@ using System;
 namespace HBSIS_Padawan.Sistema.Boletim.Models
 {
     public class Materia
-    {
-        public int Id { get; set; }
+    {        
+        public long Id { get; set; }
         public string Nome { get; set; }
-        public string Descricao { get; set; }
+        public string Descricao { get; set; }        
         public DateTime Cadastro { get; set; }
         public Status Status { get; set; }
     }
